Handled some edge cases
Note: Still can't And an empty list if using callable function.

This is because functions have no knowledge of solvers. Callable
functions access the solver through their arguments. If there aren't
any arguments passed to it, then it can't access the solver to return
a theory constant True (or other logical action for empty and).

There are two possible fixes:
1. Give functions knowledge of the solver.
	a. But it seems like if we were going to do this, we might as
	   well just make the functions solver methods...
2. Interpret python constants automatically.
	a. This way, we could just return True (the python True) from
	   the And function if its called on an emptpy list and that
	   True can be interpreted later on.